+++ src/elements/Product.php
@@ -889,6 +889,15 @@ class Product extends Element
             $record->defaultWidth = $defaultVariant->width ?? 0;
             $record->defaultWeight = $defaultVariant->weight ?? 0;
 
+            // Make sure to update the object
+            $this->defaultVariantId = $defaultVariant->id ?? null;
+            $this->defaultSku = $defaultVariant->skuAsText ?? '';
+            $this->defaultPrice = $defaultVariant->price ?? 0.0;
+            $this->defaultHeight = $defaultVariant->height ?? 0;
+            $this->defaultLength = $defaultVariant->length ?? 0;
+            $this->defaultWidth = $defaultVariant->width ?? 0;
+            $this->defaultWeight = $defaultVariant->weight ?? 0;
+
             // We want to always have the same date as the element table, based on the logic for updating these in the element service i.e resaving
             $record->dateUpdated = $this->dateUpdated;
             $record->dateCreated = $this->dateCreated;