+++ src/services/TaxRates.php
@@ -226,7 +226,6 @@ class TaxRates extends Component
                 'dateUpdated',
                 'id',
                 'include',
-                'isVat',
                 'name',
                 'rate',
                 'removeIncluded',